First slice of csv upload of phone numbers for sending messages.
At the moment the file contents are not persisted by checked in memory.
The first and last record are show if all are valid.

If there are invalid rows, they are reported and the user is prompted
to go back and sort out the upload file.

The storing of upload result (i.e. validation of file) in session will
be removed in next story which is about persisting of file for offline
processing.