Add test for template page repopulation
A bug was found whereby the body of a template was not being shown in the
on the page when returning to edit an existing template. This bug was caused
by renaming the field in some places, but not in the `Form` class.

This bug has since been fixed, but this commit adds a test to make sure that it
doesn’t happen again.