Remove ‘BETA’ label
There’s an argument to be made that the beta label should only apply to
the public-facing parts of services/platforms.

Removing it from here means that we don’t need to deal with the hassle of how it
fits into the navigation. If we have to bring it back then we could follow the
banner example here instead:

http://govuk-elements.herokuapp.com/alpha-beta-banners/#beta-banner