Move ‘new thing’ buttons alongside heading
On the send messages and manage team pages we have big green buttons for
adding/inviting a new template or team member.

On the add template page it was at the bottom, and often got missed.

On the manage team page it was at the top, but maybe too prominent because it’s
big and green.

This commit tries putting it in the top right of the page instead (except when
the template page is empty, in which case it’s unchanged).