Increase number of rows shown
We probably shouldn’t hide the contents of the CSV when people are just
testing the app, or if they’re starting off with small jobs.

A limit of 15 rows displayed was awkwardly on the cusp between just
testing and sending a small batch.

This commit increases the limit to 50; I reckon that over 50 recipients
no-one will be wanting to check them all individually.