Make border width of field layers match textbox
We put a border on text boxes that have errors. This means that they
take up more space than the layers above them which don’t have borders.

This makes the layers that highlight the fields in a template misalign.

This commit adds a line to make the borders match, which fixes this.