Use the API key pattern for showing service ID
When we show an identifier, like an ID or a key, we have a pattern which
adds a ‘copy to clipboard’ button. We weren’t using this pattern for
service ID.

This commit also moves it under the table, so hopefully people will be
less likely to confuse the service ID for an API key when scanning down
the page.