Remove activity from the nav
We saw lots of people in the lab clicking activity hoping for…
something. But it’s not really where you go to do a thing, so they
weren’t finding what they were looking for.

Since you can now get to the activity from the dashboard, let’s remove
the link in the nav, to make thing less ambiguous.