Update code to use lasest flask login.
In the process it seemed to me that the remember me checks weren't
quite right but was hidden by bug in is_authenticated in our User
model. That model returned False if login was not fresh.

However I think that should always return true for a User model as it
returns false for AnonymousUser objects.

To check for remember me instead check for login not being fresh
(which means in the current session user was loaded from cookie)

Additionally check for remember me in session being 'set'.

Changed test for remembered login as that test logged in user
properly but only session login fresh to false. That was
not correct. To mimic real behaviour the remember me cookie would
be set and notify_admin_remember_me would contain user id.