Convert uploaded Excel files to CSV
We require users to export their spreadsheets as CSV files before
uploading them. But this seems like the sort of thing a computer should
be able to do.

So this commit adds a wrapper which:
- takes a .xlsx or .xls file
- uses an appropriate library to parse the proprietary soup
- gives the data back in CSV format

We can upload the resultant CSV just like normal, and process it for
errors like normal.

Important notes
---

1. It may be worth looking at https://github.com/pyexcel/pyexcel
   (advertises that it can handle lots of formats, and detect them
   automatically)

2. This is a bit of a spike. There are no unit tests. Merge at your own
   risk, I’ll be in 🇮🇹

3. This doesn’t change the UI (ie we’re not telling anyone we can do
   this yet)