Remove exception catching on file upload
There shouldn’t be a case where we see a `ValueError` on upload any
more. Our file handling should be robust enough to deal with whatever is
thrown at it.

If it isn’t then we should raise a 500, so that we’re aware of when
users are seeing errors with their files.